# api/openapi.yaml
openapi: 3.0.3
info:
  title: SignalWatch API
  version: 2.5.0-credentials
  description: |
    SignalWatch M4 HTTP API contract. All timestamps are UTC RFC3339 values.
    The public API covers identity, profile settings, the controlled source
    catalog, subscription CRUD, and the current user's matched arXiv papers.
    Background local matching provides explainable match reasons. M4 adds the
    subscription-scoped daily Digest (scheduled in the user timezone) and email delivery Worker; mail delivery operations
    are intentionally not exposed as public endpoints. User-supplied provider credentials provide optional paper summaries and cached Digest enrichment; Bounded Agents provide persistent subscription drafting and evidence-based paper conversations. Read-only operations endpoints expose safe runtime
    status to database-authorized operator accounts; no administrative mutation
    or log-download endpoint is exposed.
...
security:
  - bearerAuth: []
paths:
...
  /api/v2/agent/conversations:
    get:
      operationId: listAgentConversations
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - items
                - page
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Conversation'
                  page:
                    type: integer
        default:
          $ref: '#/components/responses/AIError'
      parameters:
      - name: kind
        in: query
        schema:
          type: string
          enum:
          - subscription
          - paper
      - name: paper_id
        in: query
        schema:
          type: integer
          minimum: 1
      - name: page
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100000
    post:
      operationId: createAgentConversation
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        default:
          $ref: '#/components/responses/AIError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
              - kind
              properties:
                kind:
                  type: string
                  enum:
                  - subscription
                  - paper
                paper_id:
                  type: integer
                  minimum: 1
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: signalwatch_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 access token returned by the login endpoint.
...
  schemas:
    Conversation:
      type: object
      additionalProperties: false
      required:
      - id
      - kind
      - title
      - created_at
      - updated_at
      properties:
        id:
          type: string
        kind:
          type: string
          enum:
          - subscription
          - paper
        paper_id:
          type: integer
          minimum: 1
        title:
          type: string
        active_run_id:
          type: string
        latest_draft_id:
          type: string
        created_at: &id001
          $ref: '#/components/schemas/UTCTimestamp'
        updated_at: *id001
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        request_id:
          type: string
          minLength: 1
        call_id: {type: string, maxLength: 64}
        retry_after_seconds: {type: integer, minimum: 0}
        reset_at: {$ref: "#/components/schemas/UTCTimestamp"}
      example:
        code: VALIDATION_ERROR
        message: request is invalid
        request_id: req_01j_example
...
    UTCTimestamp:
      type: string
      format: date-time
      description: UTC RFC3339 timestamp.
      example: "2026-09-05T08:30:00Z"
...
  responses:
    AIError:
      description: Structured AI error. 409 denotes conflict or missing configuration; 422 invalid credentials/model; 429 local daily limit, admission interval or provider rate limit; 502 provider failure; 504 provider timeout; 503 application/storage unavailable. Retry-After is returned when a retry time is known. No model call is retried automatically.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorResponse"}
...